Add orientation option to set layout of image
This allows tiling, centering, and more control of the image,
as well as adding support for dynamic wallpapers. It also affects
pretty much every part of the file so while I could try to split
that into it's own commit this is such a large commit I wouldn't
be sure where exactly to split it without significant time dump.